PDF-566 checked core code base with -Xlint enabled to show various parser issues and applied fixes.

+++ icepdf/core/src/org/icepdf/core/io/SeekableInputConstrainedWrapper.java
@@ -102,7 +102,7 @@ public class SeekableInputConstrainedWrapper extends InputStream {
         long remain = getBytesRemaining();
         if (remain <= 0)
             return -1;
-        n = (int) Math.min(Math.min(remain, (long) n), (long) Integer.MAX_VALUE);
+        n = (int) Math.min(Math.min(remain, n), (long) Integer.MAX_VALUE);
         return streamDataInput.skip(n);
     }
